Move bad character filter to its own package
